Update to Shibboleth 3.3.0 with a 9.3 series Jetty
Revised tagging the resulting image so that the version is that of
the underlying Jetty release, rather than just "latest".

Centralise knowledge of version numbers for components in a new
VERSIONS file.

As 3.3.0 unbundles the jetty-base, fetch that as well and
infiltrate it into the Shibboleth distribution before calling
the installer.